Replace Core.Command with Cmdliner
Summary:
Pyre uses Core.Command to describe the CLI interface of the OCaml binary. The
API to define the interface is part of Core, but Core_unix is required to
actually use the specification and present a CLI interface.

Core_unix does not support Windows, so effectively Core.Command also does not
support Windows. I experimented with creating an interface to use Core.Command
on Windows, but it did not feel like a good choice, for three reasons:

1. You need to depend on "private" / internal parts of Core, like Core_unix
does, but that code is likely to change between versions. This would mean
additional maintenance.

2. The Pyre CLI is very simple. There are a handful of commands, and each
command only takes a single argument.

3. The Pyre OCaml binary does not even present the user-facing CLI. The
user-facing CLI comes from the Python frontend.

This diff replaces the use of Core.Command with a different library, Cmdliner,
which has a similar API and supports Windows. Cmdliner is used by opam and other
projects and seems like a well-maintained alternative.

While this diff preserves the CLI arguments, so existing invocations of the Pyre
CLI will continue to work, it does slightly change the `--help` output. However,
since the OCaml binary is not actually user-facing, this should be fine.

Reviewed By: rzabarazesh

Differential Revision: D49379280

fbshipit-source-id: 03a06f9cb5c07b94c063d46bfebc6b78f10a6da5