Move command initialization out of top-level evaluation
Summary:
Before this diff, commands are initialized as part of the OCaml module's
entrypoint. Entrypoint initialization is fine, but I personally find it a bit
difficult to understand when code executes, because it depends on when modules
are referenced in the source code.

This diff takes the module-static definitions for commands and makes them
functions. This makes it a bit easier to see when this code runs, but in
practice shouldn't change anything, since we still call this code as part of the
Main module's entrypoint.

Reviewed By: connernilsen

Differential Revision: D49376838

fbshipit-source-id: 441f9c28b84e4f94cba1f51a85e6f9827adf5b5a