use default `ArtifactPath` if not found in build system
Summary:
See the previous diff for more background.

This diff changes the `requestHandler` `get_modules` method to try finding modules with a default `ArtifactPath` in case the build system does not find anything.

This lets the code navigation server open files unknown to the build system and resolve imports unknown to the build system.

Some potential points of confusion:
- In a buck-built project, pyre's typechecker will _not_ find modules that are not referenced in buck targets. However, following this change, the code navigation server will. This is correct behavior but could cause confusion when users see a typechecking error but are able to go-to definition and find the definition.
- All modules within `search_paths` will now be able to be `open`ed via the code navigation server. This seems to change the meaning of `search_paths` as historically it has been distinct from `source_directories` in its inability to initiate typechecking/analysis.

Reviewed By: jasontatton

Differential Revision: D49383381

fbshipit-source-id: bf4d558c6fd2c3ccc3673fd3b985998045aa63b3