Remove dependency on Core_unix from the "log" target
Summary:
This diff removes the last caller of Core_unix, replacing it with the equivalent
call to the Unix module.

Removing the Core_unix dependency has an interesting effect: If you open Core
and use the Unix module, it errors and says to use Core_unix or Caml_unix. But
Caml_unix is defined in Core_kernel and Core_kernel is only a dependency of
Core_unix, not Core. So, you can't use Caml_unix unless you use Core_unix.

The Caml_unix from Core_kernel just re-exports Unix, so I updated this diff to
un-shadow the Unix module after opening Core.

Reviewed By: alexkassil

Differential Revision: D49397338

fbshipit-source-id: 04625063b3f5f2dc7cf5202f0871057e8b5b92e3