refactor resolve_attributes_for_name
Summary:
Background:
This stack will implement autocomplete in the code navigation daemon starting with some changes to attribute access autocomplete.

`resolve_attributes_for_name` doesn't actually take in a `Name`, instead it takes an expression. It also does not find the attributes for that expression, but the expression's parent.

This refactor changes the name of the function and calls it with the expression we want attributes from. This moves the filtering logic to the `resolve_completions_for_symbol` function.

Reviewed By: connernilsen

Differential Revision: D49501271

fbshipit-source-id: 8da4d8db915970cbcbf877da7119736b0b371dff