support transitive attributes
Summary:
Background:
This stack will implement autocomplete in the code navigation daemon starting with some changes to attribute access autocomplete.

It turns out the current implementation of autocomplete does not find attributes in any base classes. This is because we only look in the `ClassSummary` when looking for attributes. Instead, we should grab all attributes from `GlobalResolution.attributes` which allows for transitive attributes.

Because this changes the data structures (subject to change if we don't want to expose these APIs) which makes this diff big, I have left the `CompletionItemKind` to a future diff. This does temporarily break the completion symbols in VSCode but completions are disabled and I will land this as a stack.

Reviewed By: connernilsen

Differential Revision: D49468058

fbshipit-source-id: 3d0de97c09ca6f5c991d4bf6bd7e51d9791d3766