expose attribute_details
Summary:
`attribute_details` shows the attributes with some details to clients of globalResolution. This exposes a smaller API to callers than `attributes` would since that relies on implementation details of attributes.

This will primarily be used for autocomplete and will include everything necessary for autocomplete results. It can be augmented to support other queries such as signature help in the future.

Reviewed By: connernilsen

Differential Revision: D49566219

fbshipit-source-id: 2863a650b462d3c73d1e3f187b258e629a927755