Demonstrate janky check for nested class.
Summary:
Problem: When we run pyre infer, we get the infer output from the OCaml server and generate stubs in a types/ directory or apply the stubs directly to the source file.

However, we have very janky code to ignore infer output for nested classes:

https://www.internalfb.com/code/fbsource/[44cec0d0d606fedc716763ddefa9b0fca3cf276d]/fbcode/tools/pyre/client/commands/infer.py?lines=559-589

This works ok when we run pyre infer from fbcode. But, if we run it from a local root, it ends up considering all classes as nested classes.

Filed T164419913.

I added a test with an `extra_module`. It's the same for a local path, but to demonstrate that I'd have to change the test helper. Giving up for now... The task has the end-to-end repro.

Reviewed By: kinto0

Differential Revision: D49655297

fbshipit-source-id: dd82d352c3a09974b04374a2bc52082fbaea3115