do not filter results
Summary:
There are cases where autocomplete is requested in the middle of a symbol. Consider the following example from Pyls:
https://pxl.cl/3sqzW

attribute is not a prefix of attr1 or attr2 but attr (the cursor position) is.

Pyre's curent behavior filters results by the entirety of the symbol. In that example, pyre would return empty.

My attempts to filter by cursor position (D49651735) are unsuccessful because Pyre doesn't actually store the location of the identifier. We only have the start location of the `Attribute` variant which has an arbitrary expression as `base`. Any length determining logic there would be brittle incomplete.

Luckily, VSCode filters the results. This diff removes our backend filtering alltogether and relies on VSCode's filters which provides correct results for the case listed above.

Downsides:
This may have latency downsides. If performance is an issue, we may have to have to revert to our previous almost-complete solution.

Reviewed By: inseokhwang

Differential Revision: D49653520

fbshipit-source-id: 14715a41a06f57a6676a0dacb940aebd7b4fbdd7