Remove the dependency of the `Test` module on `Taint`
Summary:
The only reason we were depending on Taint was to invalidate the class definition cache on calls to `mock_scheduler`.
That is very hacky, and there is no reason why this needs to happen in `mock_scheduler`.
Let's just invalidate that cache explicitly before calls to `ModelParser.parse`. We can improve this in the future.

Reviewed By: kinto0

Differential Revision: D49727830

fbshipit-source-id: f3f3a6171eca251c0987b9b339a152b80a8c0e41