Rename `list.py` to `lists.py` to avoid typing stub collision and add types
Summary:
There was a typing stub collision for this file with the passed in `list.pyi` file in our integration test setup, here's an example: D49713814. Let's change the name of this to avoid that.

I also added types.

A good todo (T165055953) would be to error more loudly when that happens. The error only happens when a function is used by another function, and all these tests just defined functions that weren't called.

Reviewed By: arthaud

Differential Revision: D49713403

fbshipit-source-id: 33ddfe7e725ad60e39012d151ce8fae3c74df4b0