Back out "Back out "[pyre] Replace Core.Command with Cmdliner""
Summary:
The original diff caused an issue because it broke support for the `-version`
flag with a single dash. The previous diff in this stack fixes the issue by
changing to `--version` instead, so we can reland this diff as-is.

Original commit changeset: 7b81db779b42

Original Phabricator Diff: D49541875

Reviewed By: arthaud

Differential Revision: D49602403

fbshipit-source-id: 9a3c361326712ccc70c89ef836f25eceabd5cc45