Refactor `BigList` to better typing safety and `saveLoadSharedMemory` to not use big lists.
Summary:
# Refactor BigList

A better BigList api is to have it take one regular list and one big list for the append operation. That way we always append a smaller list to the bigger list.

More Context: D48577125

# saveLoadSharedMemory to not use BigList

The gains of using BigList is for O(1) length checks. This was used in the old `merge` function to guarantee that a smaller BigList was being appended to the larger one.

For the uses of `merge_same_handle` we don't need this. Specifically, we use `merge_same_handle` in two places, `CallGraph` and `GlobalConstants`, as the `merge` function to `Scheduler.map_reduce`'s named parameter `reduce`.

The type of `Scheduler.map_reduce`:
```
Scheduler.t ->
policy:Scheduler.Policy.t ->
initial:'state ->
map:('input list -> 'intermediate) ->
reduce:('intermediate -> 'state -> 'state) ->
inputs:'input list -> unit -> 'state
```

If you notice from the reduce function, the type information guarantees we are doing a left fold, and accumulating the growing list, and appending to the front of it new values. Assuming each worker produces a similarly sized list, **List.append is just as efficient as BigList.append**, and both are O(n) where n is the resulting size of the `Handle.keys`.

That is why the performance of this diff is stable, and we don't need `BigList` for this example.

Let's keep BigList around for a future usecase.

Thanks Tianhan for the nice discussion around this.

Reviewed By: tianhan0

Differential Revision: D49718741

fbshipit-source-id: e7e77a75a1b1642c3dfb4587a4eea32acf325dae