Change remove/remove_if_exists to unlink_if_exists
Summary:
This is potentially a breaking change, but auditing the callers I think it should be fine.

Sys_unix.remove is the same as the Stdlib's Sys.remove, which is itself essentially "unlink" (with some special logic for Windows around the edge case of unlinking a symlink to a directory).

Core_unix.remove is different. It's a wrapper around the C standard "remove" function, which is specified as "unlink for files, rmdir for directories."

Confusingly, PyrePath.remove used Sys_unix.remove while PyrePath.remove_if_exists used Core_unix.remove, and thus have different behavior around directories.

Furthermore, PyrePath.remove_if_exists doesn't handle a race condition, where the file is removed between the file_exists check and the remove call. If this happens, the remove call would throw and the exception would bubble up.

This diff changes all callers to a new function, unlink_if_exists. This works only on files, but existing callers only ever remove files, so this is okay. Using the Unix module directly also lets us catch the specific errors ENOENT and ENOTDIR, so we can avoid the race condition mentioned above as well.

Before, the remove function caught and logged any error, but in this diff we will actually raise those exceptions and they will bubble up. This might cause unhandled exceptions to crash the server, but this was already the case for remove_if_exists.

Reviewed By: grievejia

Differential Revision: D49704369

fbshipit-source-id: 362f701bcfea8ef5796860ba9f6113ff3e7bf155