Rename is_directory -> directory_exists
Summary:
Renaming this for clarity. If no file or directory exists at the given path,
this function returns false. An implementation based on stat + S_DIR would maybe
raise ENOENT or ENOTDIR.

This name also mirrors the existing `file_exists` function, and makes it clear
that you don't need to check both (although this doesn't happen in practice).

In later diffs, I will change the implementation of this function. The name
makes the contract of that implementation a bit clearer.

Reviewed By: grievejia

Differential Revision: D49704372

fbshipit-source-id: 11ea298eb682522be9334dfaa80e93c252836a73