Create an utility module for invoking libcst + function for finding references in a given code.
Summary:
We're going to be using libcst for few more things, so let's create its own module for libcst utility functions/classes.

Currently all it handles is parsing in code and constructing various tables of lines of symbols, symbol positions, and symbol name in various combination for a reference look up.

Reviewed By: grievejia

Differential Revision: D48751087

fbshipit-source-id: 2f880b85917c985d097744ae278a0dbaa80d04ed