Add libcst and find references
Summary:
# Context:
We want to set up Rename Symbols into few steps:
1. Enable textDocument/rename
2. Create constructs for Rename request/params/response/etc
3. Add logic to wrap references into rename response (test by hooking up references from Glean)
4. Add alternative methods to finding local references
5. Merge references from both local and global.

This diff addresses Step 4

# Changes:
Add a new file to abstract libCST logic.
Write a libCST visitor to traverse and generate a mapping of all the symbols and its references.
Helper method on the visitor to fetch the references

One thing to note is that I'm opting to save all the qualified names and their locations for now. I feel like I could reuse this for other applications, though if you rename something, you'd have to reconstruct the tree/mapping (this brings the question: should I be transformating instead of just visiting?). I could parse on document open and allow the mapping for local-find-references/go-to-def.

Reviewed By: kinto0

Differential Revision: D48158946

fbshipit-source-id: 5d69277701f94432145a9ef8451b939439d130ba