Remove core_unix dependency from Memory module
Summary:
The only complex thing here is the change to `run_tar`. Before, we used
Core_unix.create_process, which handles creating the stdin/stdout/stderr pipes
for us. Using the stdlib Unix module, we need to create these pipes ourselves.

We create a pipe for each of stdin, stdout, and stderr. Each pipe has read side
and a write side. We pass the appropriate pipes to the child (and close them on
the parent). The pipes are close-on-exec, but the pipes we pass to the child
will be `dup`'ed.

After tar exits, we also close the parent's side of the pipe. Before, this was
missing, and I think Pyre was leaking those 3 file descriptors. I'm sure this
was harmless, since this code is called at most once.

Reviewed By: arthaud

Differential Revision: D49805382

fbshipit-source-id: 60fce1180c223a84ae1a4e184625aa3cbfe504af