Reimplement class interval without using core_unix.interval_lib
Summary:
The core_unix library does not support building on Windows, so we need to break
the dependency. Unfortunately, the interval lib is in core_unix instead of core,
even for the int variant.

Happily, the interval data structure is trivial. The meet/join operations have
straightforward implementations. The new `equal` implementation is quite a bit
nicer, since we have access to the underlying structure.

Reviewed By: arthaud

Differential Revision: D49805172

fbshipit-source-id: 93a2f63395a2c2bff211681b26bf8d578667ee29