Change GetTypeErrors response signature for Pyre daemon
Summary:
This diff updates the signature of a `GetTypeErrors` response for the code navigation functionality of the Pyre daemon.

The Pyre language server response parser expects responses to be of the form `[<response_name>, <response_value>]`, where `response_value` is a JSON object that can be deserialized into a Python class. Currently, all other codenav response types have a signature that return an OCaml record (with a value containing a list type, like `Hover of { contents: HoverContent.t list }`) instead of a top-level list (`TypeErrors of Analysis.AnalysisError.Instantiated.t list`).

To allow this to be easily parsed by the Pyre language server, the signature is changed to have a `TypeErrors` response be a `{ errors: Analysis.AnalysisError.Instantiated.t list }`.

Reviewed By: grievejia

Differential Revision: D49833516

fbshipit-source-id: a28d879eb22ad1b0bbc228884399bbdf97dae2fe