Add test demonstrating nonlocal doesn't get added to call graph
Summary:
From discussion with the Pyre team, and reading pyre code, it turns out that the `captures` parameter in the define signature won't have all the data we need with regards to accesses to nonlocal variables, so we have to find a better solution.

Tentative plan:

1. Add `nonlocal_targets` information to call graph, just like we have for `global_targets`
2. Use `nonlocal_targets` for the write path of nonlocal side effects
3. Use `global_targets` for the read path of global side effects
4. Use `global_targets` for the write path of global side effects

Steps 2-4 might be gated behind a flag, depending on performance impact, especially 3 & 4.

This diff is the first step to achieving 1, by adding a call target test with `nonlocal`

Reviewed By: arthaud

Differential Revision: D49943722

fbshipit-source-id: e984623f0f877b71f70145ebbb368705dbed60b1