Fix nonlocal qualification
Summary:
Semantically, nonlocal and global qualification is very similar, because nonlocal and global are very similar.

Both search for binding in the parent frames of a function. `nonlocal` looks for variables in all but the global frame, `global` looks for variables in none but the global frame.

Context: https://peps.python.org/pep-3104/ pep that introduced `nonlocal`

This is a general fix that should help pyre, and unblocks pysa work.

Reviewed By: grievejia

Differential Revision: D49932308

fbshipit-source-id: 31d4ef5fa66e500e2fd88a259ee9dbb470a7b2ec