Add GetTypeErrors request and reponse dataclasses for codenav
Summary:
This diff adds the `GetTypeErrors` request and response dataclasses that will be used in upcoming diffs to request type errors from the Pyre daemon.

**Diff Stack Info**
This diff stack will be implementing (a gatekept) version of global, lazy, configurationless type checking to see if it improves user experiences for users with the worst type check projects.

Reviewed By: inseokhwang

Differential Revision: D49924065

fbshipit-source-id: 305aa7ff1ec048f9c9519894461ac2a1ffaa2cec