Implement get_type_errors for CodeNavigationDaemonQuerier
Summary:
This diff implements the `get_type_errors` function to get errors from the Pyre daemon and handle possible failures that occur.

**Diff Stack Info**
This diff stack will be implementing (a gatekept) version of global, lazy, configurationless type checking to see if it improves user experiences for users with the worst type check projects.

Reviewed By: inseokhwang

Differential Revision: D49924538

fbshipit-source-id: cb06d0925e533b3e7fc93f3b80c7e6377cfb64b0