Simplify code to un-shadow the stdlib Unix module
Summary:
Per feedback from Pradeep, the previous code to save the stdlib Unix module to a
temporary name, then restore the shadowed-by-Core module after was confusing.

This diff takes a different approach, which saves the original module to
CamlUnix, but does not then restore the "original" Unix module and just leaves
Core's override.

Reviewed By: connernilsen

Differential Revision: D50150099

fbshipit-source-id: 1af14175138658c2ef0841087758d606c3832950