Move type checking telemetry out of didSave and didChange telemetry events
Summary: This diff pulls type check operations outside of the timers for telemetry events on `didChange` and `didSave`. IMO, `didChange` and `didSave` are independent of the amount of time spent type checking, especially if we end up adding a debounce, which would increase the amount of time taken on these operations by a constant amount.

Reviewed By: inseokhwang

Differential Revision: D50193270

fbshipit-source-id: 90a644fbbd4774445639584ae5a464b6bffed3f4