Remove core_unix from statistics.ml
Summary:
Core_unix doesn't support Windows, so we're removing our dependency and
replacing with calls into the OCaml stdlib Unix module.

The main change here is getting the username in the statistics module. The
Core_unix getlogin function is not the same as OCaml's Unix.getlogin. Core_unix
uses getpwuid under the hood.

So, this diff replaces Core_unix.getlogin with CamlUnix.getpwuid. However,
getpwuid is not implemented on Winwdows and always raises Not_found, so we fall
back to getlogin.

On dev and corp machines, the $USER environment variable is set anyway. Flow and
Hack both have special logic to deal with OnDemand pre-warming, where the server
starts before a user has reserved the machine. Not sure if we want that for Pyre
as well.

Reviewed By: grievejia

Differential Revision: D50289710

fbshipit-source-id: f774a8b06e13a0656a69d33f0c1543e9fc43b487