Drop type errors request if daemon is busy building
Summary:
This diff blocks sending `GetTypeErrors` requests to the Pyre daemon if the daemon is not in a ready state. This prevents a bunch of overlay updates being sent with `GetTypeErrors` requests in sequence, which can cause the Pyre language server to slow down significantly in returning type errors or go to definition requests to VSCode.

**The Problem**
TL;DR: see the test case changed for an example of before/after.

Currently, on a `didChange` request, the language server stores the most recent/up to date changes in a local file cache. At the end of the `didChange` request, we send a `GetTypeErrors` request to the Pyre daemon with the most recent changes, which causes a rebuild, and eventually returns the results to the language server.

If the user makes a bunch of changes over a period of 30 seconds (giving enough time to pass VSCode's debounce, but not long enough for the Pyre daemon to complete a full build), then a long chain of rebuilds and updates occurs (since rebuilds block all other requests). Because of how long it takes for the daemon to rebuild the errors environment, this will significantly slow down results from the Pyre daemon (20+ seconds for type errors on a project that usually gets results in ~2 seconds).

**Side Note**
TL;DR: in the test case, we only send the first overlay update, but not the last one, meaning VSCode could be left in an old state. Changes will follow to fix that.

This introduces a new error however, which is that only processing `didChange` when the server is ready might leave the overlay in an invalid state or not include the most recent changes. A follow up to this diff will be made to prevent incorrect or no errors from being returned from the daemon when there are obvious errors that should be shown.

In the change to the example above, we're only processing the first `didChange` request, but any others that come along before the daemon is done building will be dropped. If the user stops typing before the daemon is ready for the next type errors request, then whatever overlay the last type errors request had will be what the daemon has access to to build the errors environment for VSCode.

Reviewed By: grievejia

Differential Revision: D50237754

fbshipit-source-id: 6072ed1d89de6c9cf6d2d38f0373fcc99bf55c26