Show Better Statuses
Summary:
## Background
Pyre statuses are shown in a box on the bottom of fbcode to display information about pyre's daemon to users. This info is currently shown only when python files are open or when the ISL is open (all of them are shown there). For that reason there's no overlap with files in other languages that may be open.

## Problem
Pyre statuses are confusing. Currently, both codenav + classic pyre show "Pyre Starting..." while starting and switch to specific messages once loaded. The specific messages are not mutually exclusive: Pyre Ready is for typechecking and CodeNav Ready is for language services. CodeNav makes no sense to anyone except members of our team.

## Solution
This diff shows clearer titles such as "Type Checking" vs "Language Services" and display these throughout the entire status bar lifetime. We leave the same hover information that specifies this is pyre.

Reviewed By: grievejia

Differential Revision: D50430812

fbshipit-source-id: bfbfa142466be6e3492d29da7502f730ac5402c1