Remove references where text does not match the original symbol names
Summary:
Context:
There can be a desync between local and Glean's copy of the code for various reasons:
1. Glean db is stale
2. User is extremely far behind Glean's latest run
3. User has already made local changes that shifts the text.
4. Many more cases.

Whenever there are situations like these, I think it'll be even less conducive for users to batch rename symbol if the user has to go back and correct the output.

Changes:
As a stop gap, we're going remove these references for now. On a later diff, I'll create a notification for users that there has been some references excluded due to text shifting.

Reviewed By: kinto0

Differential Revision: D49357912

fbshipit-source-id: 95f3b81ec434901621f180fb2e670a4d5f57164c