Call the correct querier for each LSP events in PyreLanguageServer
Summary:
Context:
We used to wrap queriers on top of each other, so `RemoteIndexDaemonQuerier` used to be responsible for checking the `CodeNavigationDaemonQuerier`'s state and determine the source of responses.

With changes from D49323317, now we have to explicitly decide in `PyreLanguageServer` on which querier is responsible for the response. This has to be checked in along with D49323317.

Changes:
Updated `PyreLanguageServer` to explicitly call `querier` vs `index_querier`
Left 2 TODO's on where we need to bubble up the state determination to follow up on another diff.

Reviewed By: grievejia

Differential Revision: D49340790

fbshipit-source-id: 86b0e56c9caf48736b3275b9091bb1ac827c6d3d