Filter out unused suppression type errors from global lazy type check errors
Summary:
With global lazy type checking, since we're not reading local configuration files, we're going to keep the type check results operating in unsafe mode to ensure we're not annoying our unsafe users, and with the assumption that our strict mode users will be adding typing anyway.

Due to this assumption, strict mode projects that have suppressions for `Any` types or untyped functions will show an error with global lazy type checking in unsafe mode complaining that their `Any` and untyped function suppressions are unused. This diff filters those errors from the results returned to the user.

Reviewed By: grievejia

Differential Revision: D50649654

fbshipit-source-id: c2311e8270afa69850c6a4bc4024bb16c805777a