Link to the wiki for errors within entrypoints.
Summary:
Problem: When there is a readonly error within a zone entrypoint, we emit an extra line of message explaining why captured variables were treated as readonly.

However, we don't link to any docs.

Joe suggested linking to the wiki.

It can also be a shorturl. No strong opinions.

Note: This is only for errors within the top-level entrypoint function. If we want the link to show up on every readonly error, even in transitive functions, that's a different thing.

Reviewed By: grievejia

Differential Revision: D50584166

fbshipit-source-id: 0117eca975182ecc7b6a0d6cea1685c037176991