Check for nested classes using module qualifier, not path.
Summary:
## Problem

When inferring type stubs for a file, we get the raw infer output from the daemon and store that to a `types/` directory. We want to ignore nested classes, probably because we don't want the extra logic of emitting nested class stubs.

The way we check for nested classes is by converting the path `foo/bar.py` to `foo.bar` and then removing that prefix from the class qualifier `foo.bar.Baz`.

This fails when the path doesn't match the qualifier, because of base-module (`dper3/dper3`), IG-under-fbcode, etc. In that case, we ignore all classes! Even for projects without base module, `pyre infer` fails when run locally instead of from `fbcode`.

## Solution

Use the module qualifier. The previous diff stored the module qualifier. Now, we use it.

Reviewed By: stroxler

Differential Revision: D50677054

fbshipit-source-id: 3656ed50dd755f1c8ddf8b41705ba039ed2895e6