encode as strings instead of SimpleElement
Summary:
the pyre client is erroring when json encoding this new argument. see test plan for details. This diff changes the argument from a SimpleElement (can't be encoded) to a string (can be encoded).

to do this we use the [command_line_argument](https://fburl.com/code/s82mhdzk) encoding function like in other places in the file.

Reviewed By: connernilsen

Differential Revision: D50748733

fbshipit-source-id: b9955d27e5ebc22dce842301ecf7b0698626898d