add backend support for targets_fallback_sources
Summary:
This diff adds backend support for targets_fallback_sources. The intention of this field is as a fallback source for when sources are not in targets or targets are building.

With this setup, only code navigation servers with buck get this configuration option. it doesn't leak into non-buck or classic servers.

Reviewed By: grievejia

Differential Revision: D50751905

fbshipit-source-id: 3e8087d37e9842a4b9fea77eae028e48d9d36e1a