respond without buck on certain queries if daemon is waiting on buck
Summary:
### Background
Buck takes a long time to build. While we're waiting, we should be able to deliver results to users.

### Change
Instead of a lock on everything, we should only have a lock to ensure concurrent code updates. For queries where we prefer availability to consistency, we can skip this lock and start trying to look for source code to analyze.

This improves things when buck is building because of our earlier mitigation of skipping buck when we don't find any sources with it (D49383381). By using whatever is available at the time of the query to find the result, we are able to look on disk when buck is building.

See [this document](https://docs.google.com/document/d/1kqokb5DMv1XMaYG-TOqrU4gti6kqAsCbhwwv4L-1SUE/edit) for more details on the theory behind this diff and the next diff in the stack for a more in-depth test plan.

Reviewed By: grievejia

Differential Revision: D50654522

fbshipit-source-id: 4f9fb56dcb629db1a604edfa571e539b0a571b38