Fix parse of newline character
Summary:
Users have reported that newline characters are escaped in the hover overlay on pyre powered codenav:

https://pxl.cl/3H1XM

This is because strings out of errpy come with escaped newline characters as part of the ocamlrep marshalling process from Rust to OCaml (I expect there is a good reason why this is done in the Hack space).

e.g. `"\n something \n"` is represented as `"\n something \n"`

For Pyre all this means is that we need to un-escape the escaped newline characters. This is fixed here

Reviewed By: kinto0

Differential Revision: D50760264

fbshipit-source-id: 7dc90613d6afaa7bda3822524eb7b55b497184b1