Mark `dict[str, Any]` as having `Any` type.
Summary:
Problem: Pyre considers `list[Any]` and `dict[Any, Any]` as having coverage gaps. However, it specifically ignores `dict[str, Any]` because it is widespread and we were concerned that people would see lots of squiggles in the IDE.

Now that IG is pushing for 99.9% expression-coverage and wants to get rid of `dict[str, Any]` (and people no longer use the IDE squiggles for expression-coverage), it's misleading to say that only 3.5% of expressions have `Any`, since that excludes `dict[str, Any]`.

Solution: Include it in the expression-level coverage.

Note: This won't affect typechecking results. This is just for the %expression-coverage.

Reviewed By: grievejia

Differential Revision: D50709854

fbshipit-source-id: 6e4a9093358d64564f0745098fba0e65a0a98225