Restart pyre server on connectivity failure
Summary:
Presently pyre codenav does not restart the pyre server when the socket connection is deleted from the filesystem or when the pyre server hangs when pyre kill is invoked.

This is manifests as a `connections.ConnectionFailure` (there are also two other less common failures which we can restart on: `asyncio.IncompleteReadError` and `ConnectionError`). Here we add some code to boil this restartable failure as an `error_source` up to `dispatch_nonblocking_request` - where the client can choose to restart based on the presence of and type of exception raised.

Event is implicitly logged to scuba (with more detailed error message including stack trace) by previous diff in stack.

In later diffs:
 - Refactor dispatch_nonblocking_request, use match syntax etc.

Reviewed By: grievejia

Differential Revision: D50738586

fbshipit-source-id: 82754ae757291e6c2357e80ea678e8673a5781de