Remove `track_dependency` asymmetry from AstEnvironment
Summary:
The dependency tracking logic between AstEnvironment and
UnannotatedGlobalEnvironment is in principle pretty standard, but the code
itself has long been irrritating to me because instead of following the normal
rules that an environment owns creating its own dependency keys, we allowed
AstEnvironment to create the `WildcardImport` key on behalf of
UnannotatedGlobalEnvironment.

In principle there are some good reasons for this - because the key is always
the qualifier we are looking up, we can take a boolean flag instead of the full
key. But the fact that it makes the api of `AstEnvironment.ReadOnly` behave
differently than all other read-only apis means it will get in the way of
any refactors attempting a major overhaul to abstract away dependency tracking.

Let's just standardize on environments always creating their own dependency
keys; in this case that means pulling the definition of the dependency key
up into a helper function at the top of UnannotatedGlobalEnvironment.ml.

Reviewed By: grievejia

Differential Revision: D50841446

fbshipit-source-id: 04325bc4b30e008f89d177b41210d3ba5758a9e5