Do not use IndexTracker in class hierarchy
Summary:
We currently intern class names into a SharedMemory table via IndexTracker.

This isn't dependency-tracked because the mapping persists for the lifetime of
the server, but it still introduces a dependency on shared memory that we
ideally don't want to have to deal with in future versions of Pyre (e.g.
in a buck-oriented pyre, and more importantly in an Ocaml 5.x oriented Pyre
using domain-based parallelism).

The only reason to do all this is performance, but it is actually costing us RAM footprint,
and the perf affect seems negligible (ServiceLab suggests we do better without interning
in Pyre; Pysa may be just a little slower but both differences are small).

This diff only removes the use of IndexTracker because I wanted a minimal diff for
CI to give us perf numbers; I clean up all the remaining references to it in the next diff.

Reviewed By: grievejia

Differential Revision: D50887959

fbshipit-source-id: 6b86d3c6921482aa7035a935719fcb8f58363c3b